fix mouse aim and make gun rays visible again

+++ src/se/BaseUlterior/Utils/UlteriorUtils.java
@@ -28,8 +28,8 @@ public class UlteriorUtils {
 		return false;
 	}
 
-	public static int angleToPoint(float x1, float y1, float x2, float y2) {
-		return (int) Math.toDegrees(Math.atan2(y2 - y1, x2 - x1));
+	public static float angleToPoint(float x1, float y1, float x2, float y2) {
+		return (float) Math.toDegrees(Math.atan2(y2 - y1, x2 - x1));
 	}
 
 	public static int distance(GameObject go1, GameObject go2) {
